Phase 6: Extract historical mode into historical-mode.js module
- Created historical-mode.js (1,742 lines) with factory pattern
- Removed 1,742 lines from app.js (10,494 → 8,790 lines, -16%)
- Updated 35 function calls to use HistoricalMode module
- Added module loading in index.html

Optimizations included:
- Object pooling for Vector3 and Color instances
- DOM element caching for performance
- Configurable DEBUG flag for console logging
- Pre-allocated arrays in track rendering
- Cached calculations (altitude scaling)
- Proper memory cleanup in clearHistoricalTracks()

Functions extracted:
- Track rendering (5 functions)
- Heat map & corridors (5 functions)
- Playback system (10 functions)
- Controls setup (2 functions)
- Mode switching (1 function)
- Detail & filters (2 functions)

Total Phases 1-6: 3,638 lines reduced (-29.3% from original 12,428)
