+++ public/index.html
@@ -3937,6 +3937,9 @@
             // Load data-service-historical.js as a module (refactored historical data loading)
             loadScript('data-service-historical.js', true);
 
+            // Load historical-mode.js as a module (historical track rendering, playback, heat maps)
+            loadScript('historical-mode.js', true);
+
             // Load aircraft-svg-system.js
             loadScript('aircraft-svg-system.js');
